spacemacs

+++ linkAll.sh
@@ -7,10 +7,14 @@ DOT_FILES=(${DOT_FILES[@]} .gitconfig .gitignore)
 
 for file in ${DOT_FILES[@]}
 do
-  if [ -a $HOME/$file ]; then
-    echo "$file is already exist"
-    mv $HOME/$file $HOME/$file.org
-  fi
-  echo "$file link to dotfiles/$file"
-  ln -s $HOME/dotfiles/$file $HOME/$file
+    if [ -L $HOME/$file ]; then
+        echo "symbolic link $file is already exist."
+    else
+        if [ -a $HOME/$file ]; then
+            echo "$file is already exist. backup to $HOME/$file.org"
+            mv $HOME/$file $HOME/$file.org
+        fi
+        echo "$file link to dotfiles/$file"
+        ln -s $HOME/dotfiles/$file $HOME/$file
+    fi
 done
